requirements.txt
# Kintsugi AI - Requirements
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data Loading
tqdm>=4.65.0

# Web Interface
gradio>=4.0.0

# Visualization (optional)
matplotlib>=3.7.0

# Jupyter (for Colab)
ipywidgets>=8.0.0
